requirements.txt
torch
pyro-ppl
matplotlib
pyyaml
normflows @ git+https://github.com/VincentStimper/normalizing-flows.git
